backend/requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
albucore==0.0.17
albumentations==1.4.18
alembic==1.14.1
annotated-types==0.7.0
anyio==4.5.2
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.3.0
backcall==0.2.0
bcrypt==4.2.1
blinker==1.8.2
Brotli==1.1.0
cachetools==5.5.0
catalyst==20.9
causal-conv1d @ file:///root/autodl-tmp/UNetMamba-main/causal_conv1d-1.2.0.post2%2Bcu118torch2.0cxx11abiFALSE-cp38-cp38-linux_x86_64.whl#sha256=534453b2e4266f734c8fcc06338b067ac547ef962aa9e14c5f86fe5a68d5e3b8
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.2
ConfigArgParse==1.7
contourpy==1.1.1
cryptography==44.0.1
cycler==0.12.1
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
ecdsa==0.19.0
einops==0.7.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.8
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
fvcore==0.1.5.post20221221
gevent==24.2.1
geventhttpclient==2.0.12
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.37.0
google-auth-oauthlib==1.0.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
imageio==2.34.0
importlib_metadata==7.1.0
importlib_resources==6.4.5
iopath==0.1.10
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lightning==2.2.2
lightning-utilities==0.11.2
lit==18.1.8
locust==2.25.0
Mako==1.3.9
mamba-ssm @ file:///root/autodl-tmp/UNetMamba-main/mamba_ssm-1.1.1%2Bcu118torch2.0cxx11abiFALSE-cp38-cp38-linux_x86_64.whl#sha256=4a202baee9e00f850dd9c600853f56b100b0e5c3d903f6f9196170cf9e1fef0b
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mat4py==0.6.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mmdet==3.3.0
mmengine==0.10.1
mmpretrain==1.2.0
mmsegmentation==1.2.2
model-index==0.1.11
modelindex==0.0.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.1
ninja==1.11.1.3
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai==1.65.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
ordered-set==4.1.0
packaging==24.2
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==4.3.6
plotly==5.24.1
portalocker==3.0.0
prettytable==3.11.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.29.2
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycocotools==2.0.7
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytorch-lightning==2.2.2
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
roundrobin==0.0.4
rsa==4.9
safetensors==0.4.5
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
selective_scan @ file:///root/autodl-tmp/UNetMamba-main/unetmamba_model/kernels/selective_scan
shapely==2.0.6
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.44.0
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.15
tokenizers==0.19.1
tomli==2.2.1
torch @ file:///root/autodl-tmp/UNetMamba-main/torch-2.0.0%2Bcu118-cp38-cp38-linux_x86_64.whl#sha256=1f8efaebfcbb7ec3962fd8c7c3be02c6666eff53a12043006a749d647656163e
torchaudio==2.0.1
torchmetrics==1.3.2
torchvision @ file:///root/autodl-tmp/UNetMamba-main/torchvision-0.15.1%2Bcu118-cp38-cp38-linux_x86_64.whl#sha256=9104730d655abf282c2845d4cd472bb08936f3c85083a28aefd76ad9eeaff261
tqdm==4.57.0
traitlets==5.14.3
transformers==4.40.1
triton==2.0.0
ttach==0.0.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.33.0
wcwidth==0.2.13
Werkzeug==3.0.6
yacs==0.1.8
yapf==0.43.0
yarl==1.15.2
zipp==3.20.2
zope.event==5.0
zope.interface==7.2
